# Conflicts:
#	Makefile
#	fm_voice.c

+++ syx_dx21.h
@@ -217,7 +217,7 @@ int dx21_midi_receive(struct dx21_midi_receiver *rx, uint8_t byte);
 
 void dx21_vced_voice_bank_init(struct dx21_vced_voice_bank *bank);
 int dx21_vced_voice_bank_from_buffer(struct dx21_vced_voice_bank *bank, uint8_t *buf, size_t filesize);
-int dx21_vced_voice_bank_send(struct dx21_vced_voice_bank *bank, size_t (*write)(void *, size_t, void *), void *data_ptr);
+int dx21_vced_voice_bank_send(struct dx21_vced_voice_bank *bank, int (*write)(void *, size_t, void *), void *data_ptr);
 #ifndef __EMSCRIPTEN__
 void dx21_vced_voice_dump(struct dx21_vced_voice *voice, int voicenum);
 void dx21_vced_voice_bank_dump(struct dx21_vced_voice_bank *bank);
